+++ src/main/java/com/example/bak/global/common/response/FailureApiResponse.java
@@ -1,9 +1,9 @@
-package com.example.bak.global.common;
+package com.example.bak.global.common.response;
 
 import lombok.Getter;
 
 @Getter
-public class FailureApiResponse extends ApiResponse{
+public class FailureApiResponse extends ApiResponse {
 
     protected FailureApiResponse(String message) {
         super(false, message);
